discord_bot/requirements.txt
discord
python-dotenv
openai
requests
